+++ actiontrail-20200706/core/Models/DescribeTrailsResponseBody.cs
@@ -58,6 +58,14 @@ namespace AlibabaCloud.SDK.Actiontrail20200706.Models
             [Validation(Required=false)]
             public bool? IsOrganizationTrail { get; set; }
 
+            [NameInMap("MaxComputeProjectArn")]
+            [Validation(Required=false)]
+            public string MaxComputeProjectArn { get; set; }
+
+            [NameInMap("MaxComputeWriteRoleArn")]
+            [Validation(Required=false)]
+            public string MaxComputeWriteRoleArn { get; set; }
+
             /// <summary>
             /// The name of the trail.
             /// </summary>