+++ actiontrail-20200706/core/Models/GetDeliveryHistoryJobResponse.cs
@@ -10,15 +10,15 @@ namespace AlibabaCloud.SDK.Actiontrail20200706.Models
 {
     public class GetDeliveryHistoryJobResponse : TeaModel {
         [NameInMap("headers")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public Dictionary<string, string> Headers { get; set; }
 
         [NameInMap("statusCode")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public int? StatusCode { get; set; }
 
         [NameInMap("body")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public GetDeliveryHistoryJobResponseBody Body { get; set; }
 
     }