+++ actiontrail-20200706/core/Models/UpdateTrailResponseBody.cs
@@ -23,6 +23,14 @@ namespace AlibabaCloud.SDK.Actiontrail20200706.Models
         [Validation(Required=false)]
         public string HomeRegion { get; set; }
 
+        [NameInMap("MaxComputeProjectArn")]
+        [Validation(Required=false)]
+        public string MaxComputeProjectArn { get; set; }
+
+        [NameInMap("MaxComputeWriteRoleArn")]
+        [Validation(Required=false)]
+        public string MaxComputeWriteRoleArn { get; set; }
+
         /// <summary>
         /// The name of the trail.
         /// </summary>