+++ actiontrail-20200706/ChangeLog.md
@@ -1,3 +1,14 @@
+2024-02-21 Version: 2.3.0
+- Support API EnableInsight.
+- Update API CreateTrail: add param MaxComputeProjectArn.
+- Update API CreateTrail: add param MaxComputeWriteRoleArn.
+- Update API CreateTrail: update response param.
+- Update API DescribeTrails: update response param.
+- Update API UpdateTrail: add param MaxComputeProjectArn.
+- Update API UpdateTrail: add param MaxComputeWriteRoleArn.
+- Update API UpdateTrail: update response param.
+
+
 2023-11-07 Version: 2.2.0
 - Generated csharp 2020-07-06 for Actiontrail.